fix(plugin): log less verbose when a class cannot be rewritten
Usually we fail to rewrite classes when loading the class fails, this
mostly happens when the class has imports that are not on the class path
Those classes are probably not loaded at runtime (or else the app would
crash), so we don't need to rewrite them anyways. Typical use cases are
integration classes that are loaded conditionally if a peer library is
present, e.g. new relic has classes for okhttp - but only uses them if
the app actually uses okhttp.