fix: Send the same country code to the config api and in the payload
Previously, the config API was always sent the country code found from the device locale, whereas the payload was sending the SIM card's country code or the locale if no SIM card is found. This commit changes the ConfigStore to use the same method for retrieving the country code as the payload does.